replaced images with content
