+++ app/app.css
@@ -71,6 +71,9 @@
     "Inter", ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
     "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
   --font-lato: Lato;
+  --color-light: #f1f1f1;
+  --color-primary: #1a2e45;
+  --color-secondary: #e1a730;
 }
 
 html,
@@ -92,4 +95,21 @@ body {
 }
 .fixed-footer {
   @apply fixed bottom-0 left-0 w-screen z-0;
+}
+
+.underlineDecor {
+  --height: 6px;
+  --backgroundColor: white;
+  --bottom: -0.5lh;
+  position: relative;
+  &:after {
+    position: absolute;
+    content: "";
+    display: block;
+    width: calc(120rem / 16);
+    height: var(--height);
+    background-color: var(--backgroundColor);
+    left: 0;
+    bottom: var(--bottom);
+  }
 }
