+++ custom.d.ts
@@ -1,0 +1,7 @@
+import "react";
+
+declare module "react" {
+  interface CSSProperties {
+    [key: `--${string}`]: string | number;
+  }
+}
